refactor(bayn): isolate candidate failures
